+++ src/Layout.tsx
@@ -1,5 +1,5 @@
 import { lazy, Suspense } from 'react'
-import Alert from 'src/components/common/Alert'
+import { AlertCMS } from 'src/components/common/Alert'
 import Footer from 'src/components/common/Footer'
 import Navbar from 'src/components/common/Navbar'
 import Toast from 'src/components/common/Toast'
@@ -13,9 +13,7 @@ function Layout({ children }: PropsWithChildren<unknown>) {
 
   return (
     <div id="layout">
-      <Alert>
-        Get 10% off today:&nbsp;<span>NEW10</span>
-      </Alert>
+      <AlertCMS />
 
       <Navbar />